requirements.txt
# Use CUDA 12.1 wheels from PyTorch
--extra-index-url https://download.pytorch.org/whl/cu121

# ---- PyTorch (GPU) ----
torch==2.4.0+cu121
torchvision==0.19.0+cu121
pytorch-msssim==1.0.0

# ---- core scientific stack ----
numpy==1.26.4
scipy==1.11.4
scikit-image==0.22.0
tifffile==2024.8.30
pillow==10.3.0
matplotlib==3.8.4
tqdm==4.66.4
pyyaml==6.0.1
requests==2.32.3
pandas==2.3.1

# ---- vision / utils ----
opencv-python==4.10.0.84
einops==0.8.0
visdom==0.2.4
dominate==2.9.1
beautifulsoup4==4.13.4

# ---- optional ----
wandb==0.21.1   # remove this line if you don't actually use wandb
